# docs/generated/openapi.yaml
openapi: 3.0.1
info:
  title: Blog API
  description: |-
    Auto-generated OpenAPI spec via springdoc-openapi.

    - Web APIs: /api
    - Admin APIs: /api/admin

    Admin APIs require: Authorization: Bearer <token>
  version: v1
servers:
- url: http://localhost:8080
  description: Local dev
security:
- bearer-jwt: []
...
paths:
  /api/admin/tags/{id}:
    put:
      tags:
      - Admin - Tags
      summary: Update tag
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameOnlyRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
    delete:
      tags:
      - Admin - Tags
      summary: Delete tag
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /api/admin/posts/{id}:
    get:
      tags:
      - Admin - Posts
      summary: Get post detail for editing
      operationId: detail_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAdminPostEditVO'
    put:
      tags:
      - Admin - Posts
      summary: Update post
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPostUpdateRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
    delete:
      tags:
      - Admin - Posts
      summary: Delete post
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/admin/posts/{id}/unpublish:
    put:
      tags:
      - Admin - Posts
      summary: Unpublish post
      operationId: unpublish
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/admin/posts/{id}/publish:
    put:
      tags:
      - Admin - Posts
      summary: Publish post
      operationId: publish
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /api/admin/tags:
    get:
      tags:
      - Admin - Tags
      summary: List tags
      operationId: list_4
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListTagVO'
    post:
      tags:
      - Admin - Tags
      summary: Create tag
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameOnlyRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLong'
  /api/admin/posts:
    get:
      tags:
      - Admin - Posts
      summary: List posts
      description: Admin post list with pagination and filters.
      operationId: page_1
      parameters:
      - name: q
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/AdminPostQuery'
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageResultAdminPostListItemVO'
    post:
      tags:
      - Admin - Posts
      summary: Create post
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPostCreateRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLong'
...
  /api/posts/{id}:
    get:
      tags:
      - Web - Posts
      summary: Get published post detail
      description: Returns post detail and increases viewCount by 1.
      operationId: detail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePostDetailVO'
...
  /api/admin/posts/meta/tags:
    get:
      tags:
      - Admin - Posts
      summary: List tags for editor
      operationId: tags
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListTagVO'
  /api/admin/posts/meta/categories:
    get:
      tags:
      - Admin - Posts
      summary: List categories for editor
      operationId: categories
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListCategoryVO'
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: object
          description: Response data payload
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    NameOnlyRequest:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 50
          minLength: 0
          type: string
          description: Name
          example: Java
...
    AdminPostUpdateRequest:
      required:
      - content
      - title
      type: object
      properties:
        title:
          maxLength: 120
          minLength: 0
          type: string
          description: Post title
          example: Hello World
        summary:
          maxLength: 300
          minLength: 0
          type: string
          description: Post summary
          example: A short introduction about this post
        content:
          type: string
          description: Post content in Markdown
          example: |-
            # Heading

            Content...
        coverUrl:
          maxLength: 255
          minLength: 0
          type: string
          description: Cover image URL (usually under /uploads)
          example: /uploads/202602/cover.jpg
        categoryId:
          type: integer
          description: Category id
          format: int64
          example: 1
        tagIds:
          type: array
          description: Tag ids
          example:
          - 1
          - 2
          - 3
          items:
            type: integer
            description: Tag ids
            format: int64
...
    ApiResponseLong:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: integer
          description: Response data payload
          format: int64
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
...
    AdminPostCreateRequest:
      required:
      - content
      - title
      type: object
      properties:
        title:
          maxLength: 120
          minLength: 0
          type: string
          description: Post title
          example: Hello World
        summary:
          maxLength: 300
          minLength: 0
          type: string
          description: Post summary
          example: A short introduction about this post
        content:
          type: string
          description: Post content in Markdown
          example: |-
            # Heading

            Content...
        coverUrl:
          maxLength: 255
          minLength: 0
          type: string
          description: Cover image URL (usually under /uploads)
          example: /uploads/202602/cover.jpg
        categoryId:
          type: integer
          description: Category id
          format: int64
          example: 1
        tagIds:
          type: array
          description: Tag ids
          example:
          - 1
          - 2
          - 3
          items:
            type: integer
            description: Tag ids
            format: int64
...
    ApiResponseListTagVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: array
          description: Response data payload
          items:
            $ref: '#/components/schemas/TagVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    TagVO:
      type: object
      properties:
        id:
          type: integer
          description: Tag id
          format: int64
          example: 1
        name:
          type: string
          description: Tag name
          example: Spring
      description: Tag
...
    CategoryVO:
      type: object
      properties:
        id:
          type: integer
          description: Category id
          format: int64
          example: 1
        name:
          type: string
          description: Category name
          example: Java
      description: Category
...
    ApiResponsePostDetailVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          $ref: '#/components/schemas/PostDetailVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    PostDetailVO:
      type: object
      properties:
        id:
          type: integer
          description: Post id
          format: int64
          example: 1
        title:
          type: string
          description: Post title
          example: Hello World
        summary:
          type: string
          description: Post summary
          example: Short introduction
        content:
          type: string
          description: Post content in Markdown
          example: |-
            # Heading

            Content...
        coverUrl:
          type: string
          description: Cover image URL
          example: /uploads/202602/cover.jpg
        category:
          $ref: '#/components/schemas/CategoryVO'
        tags:
          type: array
          description: Tags
          items:
            $ref: '#/components/schemas/TagVO'
        publishedAt:
          type: string
          description: Published time (string)
          example: 2026-02-01 12:00:00
        viewCount:
          type: integer
          description: View count
          format: int64
          example: 123
      description: Post detail
...
    ApiResponseListCategoryVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: array
          description: Response data payload
          items:
            $ref: '#/components/schemas/CategoryVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
...
    AdminPostQuery:
      type: object
      properties:
        pageNum:
          minimum: 1
          type: integer
          description: Page number (start from 1)
          format: int64
          example: 1
        pageSize:
          maximum: 100
          minimum: 1
          type: integer
          description: Page size (1~100)
          format: int64
          example: 10
        status:
          maxLength: 20
          minLength: 0
          type: string
          description: "Post status (optional): DRAFT/PUBLISHED"
          example: PUBLISHED
          enum:
          - DRAFT
          - PUBLISHED
        keyword:
          maxLength: 50
          minLength: 0
          type: string
          description: Keyword for searching title
          example: spring
        categoryId:
          type: integer
          description: Filter by categoryId
          format: int64
          example: 1
        tagId:
          type: integer
          description: Filter by tagId
          format: int64
          example: 2
    AdminPostListItemVO:
      type: object
      properties:
        id:
          type: integer
          description: Post id
          format: int64
          example: 1
        title:
          type: string
          description: Post title
          example: Hello World
        status:
          type: string
          description: Status
          example: PUBLISHED
          enum:
          - DRAFT
          - PUBLISHED
        category:
          $ref: '#/components/schemas/CategoryVO'
        publishedAt:
          type: string
          description: Published time
          example: 2026-02-01 12:00:00
        updatedAt:
          type: string
          description: Updated time
          example: 2026-02-01 12:00:00
      description: Admin post list item
    ApiResponsePageResultAdminPostListItemVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          $ref: '#/components/schemas/PageResultAdminPostListItemVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    PageResultAdminPostListItemVO:
      type: object
      properties:
        list:
          type: array
          description: Current page list
          items:
            $ref: '#/components/schemas/AdminPostListItemVO'
        total:
          type: integer
          description: Total items count
          format: int64
          example: 100
        pageNum:
          type: integer
          description: Current page number
          format: int64
          example: 1
        pageSize:
          type: integer
          description: Page size
          format: int64
          example: 10
      description: Pagination result wrapper
    AdminPostEditVO:
      type: object
      properties:
        id:
          type: integer
          description: Post id
          format: int64
          example: 1
        title:
          type: string
          description: Title
          example: Hello World
        summary:
          type: string
          description: Summary
          example: Short introduction
        content:
          type: string
          description: Content in Markdown
          example: |-
            # Heading

            Content...
        coverUrl:
          type: string
          description: Cover image URL
          example: /uploads/202602/cover.jpg
        categoryId:
          type: integer
          description: Category id
          format: int64
          example: 1
        tagIds:
          type: array
          description: Tag ids
          example:
          - 1
          - 2
          - 3
          items:
            type: integer
            description: Tag ids
            format: int64
        status:
          type: string
          description: Status
          example: DRAFT
          enum:
          - DRAFT
          - PUBLISHED
      description: Admin post edit model
    ApiResponseAdminPostEditVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          $ref: '#/components/schemas/AdminPostEditVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
...
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT